# api-contract-v1.yaml
openapi: 3.1.0
info:
  title: Event Face Matching API
  version: 1.0.0
  description: |
    Phase 1 contract — single studio, no billing. Multi-tenant structures are
    present but only one studio row exists.

    ## Who owns what
    Backend (routes, DB, worker) is implemented by the studio owner.
    Frontend (three React apps) is implemented against this spec via a mock server.
    Neither side blocks the other. When the real server is live, only the base URL changes.

    ## Two separate auth contexts — do not mix them
    - **Staff** (studio owners, photographers, platform admins) use
      `Authorization: Bearer <access_token>`.
    - **Guests** have no account. They use `X-Guest-Session: <session_token>`,
      scoped to exactly one event, and can reach only the `/g/*` routes.

    Keeping these in different headers is deliberate: it makes it structurally
    hard to accidentally accept a guest token on a staff route.

    ## photos.status state machine
    ```
    pending ──► uploaded ──► processing ──► done
                                 │
                                 └────────► failed ──► (retry) ──► processing
    ```
    - `pending`    row created, presigned URL issued, bytes not yet confirmed
    - `uploaded`   client confirmed the PUT succeeded; job enqueued
    - `processing` a worker has claimed the job
    - `done`       faces extracted, embedded, stored, matched against guests
    - `failed`     gave up after 5 attempts; `error_code` explains why

...

...
servers:
...
security:
  - staffAuth: []
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Exchange an identifier and password for tokens
      description: |
        Accounts are created by hand, never by self-signup. The platform creates
        a studio and its owner; the owner creates their photographers. Credentials
        are delivered over WhatsApp.

        `identifier` accepts **either a username or an email address**. Many small
        Indian studios do not use email, so requiring one would exclude real
        customers, but studios that do have one should be able to use it. Match
        on username first, then on email.

        Recovery for an account with no email is a manual reset from the admin
        panel, verified against the phone number on the studio record.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier:
                  type: string
                  minLength: 3
                  maxLength: 120
                  description: Username or email address.
                password: { type: string, minLength: 8 }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Exchange a refresh token for a new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: Refreshed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/password:
    post:
      tags: [auth]
      summary: Change the signed-in user's password
      description: |
        Also clears `must_change_password`. Studios are created by the platform
        with an initial password, and this is how that gets replaced on first
        sign-in.

        All other refresh tokens for this user are revoked, so a leaked initial
        password stops working the moment the real one is set.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string }
                new_password: { type: string, minLength: 10 }
      responses:
        '200':
          description: Changed. Returns a fresh token pair.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422':
          description: New password rejected
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                error:
                  code: PASSWORD_TOO_WEAK
                  message: Use at least 10 characters.
  /me:
    get:
      tags: [auth]
      summary: Current user, their studio, and their role
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ───────────────────────────── events ─────────────────────────────
...
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke a refresh token family
      description: |
        Revokes the whole family the token belongs to, so every device in that
        chain is signed out. Never reports whether the token existed: an
        endpoint that distinguishes them is a way to test whether a stolen
        token is still live.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '204': { description: Signed out }
...
components:
  securitySchemes:
    staffAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Studio owners, photographers and platform admins.
    guestAuth:
      type: apiKey
      in: header
      name: X-Guest-Session
      description: Opaque session token, scoped to one event, valid for `/g/*` only.
...
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable identifier. Frontend switches on this, never on `message`.
            message:
              type: string
              description: Human-readable text, safe to show a user.
            details:
              type: object
              additionalProperties: true
    TokenPair:
      type: object
      required: [access_token, refresh_token, expires_in]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        expires_in: { type: integer, description: Seconds until the access token expires. }
    Me:
      type: object
      required: [user_id, username, studio, role, must_change_password]
      properties:
        user_id: { type: string, format: uuid }
        username: { type: string }
        email:
          type: [string, 'null']
          description: Optional. Present only for accounts that supplied one.
        name: { type: [string, 'null'] }
        role: { type: string, enum: [owner, photographer, platform_admin] }
        must_change_password:
          type: boolean
          description: |
            True while the account still carries the password whoever created it
            chose. The app blocks everything else until it is changed, so nobody
            keeps working knowledge of someone else's credentials.
        studio:
          type: object
          required: [id, name]
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
            phone:
              type: [string, 'null']
              description: |
                Used to confirm who is calling during a manual password reset,
                and to reach a studio on WhatsApp. Not used for OTP.
            brand_color: { type: [string, 'null'] }
            brand_logo_url: { type: [string, 'null'] }
...